Showing an empty tab for Bolt default tabs
Empty tabs are shown when you create a field with condition and group/tab titled `Taxonomy`.
For some reason there is always an empty div created with attribute `data-bolt-fieldset` for taxonomy tab when it’s used in the CT’s. It looks like it’s Bolt’s default behaviour for that tab.
So when you create `taxonomylist` type fields and set their group as `taxonomy` for pages with conditions to show them only on certain page templates, they still appear in all the templates due to that Bolt’s empty fieldset with the empty data attribute for `data-bolt-fieldset`.
The fix is to check if the data attribute is not empty when counting the visible fields length.